+++ app/src/renderer/vuex/modules/delegation.js
@@ -166,7 +166,7 @@ export default ({ node }) => {
         dispatch,
         commit
       },
-      { stakingTransactions }
+      { stakingTransactions, password }
     ) {
       const denom = config.bondingDenom.toLowerCase()
       const delegatorAddr = wallet.address
@@ -213,7 +213,8 @@ export default ({ node }) => {
         to: wallet.address, // TODO strange syntax
         delegations: mappedDelegations,
         begin_unbondings: mappedUnbondings,
-        begin_redelegates: mappedRedelegations
+        begin_redelegates: mappedRedelegations,
+        password
       })
 
       if (mappedDelegations) {