+++ test/unit/specs/components/staking/DelegationModal.spec.js
@@ -67,6 +67,23 @@ describe(`DelegationModal`, () => {
         `cosmosvaladdr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctplpn3au`
       )
     })
+
+    it(`account password defaults to an empty string`, () => {
+      expect(wrapper.vm.password).toEqual(``)
+    })
+
+    it(`password is hidden by default`, () => {
+      expect(wrapper.vm.showPassword).toBe(false)
+    })
+  })
+
+  describe(`Password display`, () => {
+    it(`toggles the password between text and password`, () => {
+      wrapper.vm.togglePassword()
+      expect(wrapper.vm.showPassword).toBe(true)
+      wrapper.vm.togglePassword()
+      expect(wrapper.vm.showPassword).toBe(false)
+    })
   })
 
   describe(`enables or disables the Delegation button correctly`, () => {
@@ -77,6 +94,7 @@ describe(`DelegationModal`, () => {
       })
 
       it(`if the user manually inputs a number greater than the balance`, () => {
+        wrapper.setData({ password: `1234567890` })
         let amountField = wrapper.find(`#amount`)
         amountField.element.value = 142
 
@@ -86,11 +104,17 @@ describe(`DelegationModal`, () => {
         amountField.trigger(`input`)
         expect(amountField.element.value).toBe(`100`)
       })
+
+      it(`if the password field is empty`, () => {
+        wrapper.setData({ amount: 10, password: `` })
+        let delegationBtn = wrapper.find(`#submit-delegation`)
+        expect(delegationBtn.html()).toContain(`disabled="disabled"`)
+      })
     })
 
     describe(`enables the 'Delegation' button`, () => {
       it(`if the amout is positive and the user has enough balance`, () => {
-        wrapper.setData({ amount: 50 })
+        wrapper.setData({ amount: 50, password: `1234567890` })
 
         let delegationBtn = wrapper.find(`#submit-delegation`)
         expect(delegationBtn.html()).not.toContain(`disabled="disabled"`)
@@ -100,7 +124,7 @@ describe(`DelegationModal`, () => {
 
   describe(`(Re)delegate`, () => {
     it(`Delegation button submits a (re)delegation and closes modal`, () => {
-      wrapper.setData({ amount: 50 })
+      wrapper.setData({ amount: 50, password: `1234567890` })
       wrapper.vm.onDelegation()
 
       expect(wrapper.emittedByOrder()).toEqual([
@@ -109,7 +133,8 @@ describe(`DelegationModal`, () => {
           args: [
             {
               amount: 50,
-              from: `cosmos15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9`
+              from: `cosmos15ky9du8a2wlstz6fpx3p4mqpjyrm5ctpesxxn9`,
+              password: `1234567890`
             }
           ]
         },